refactor(collection): extract text normalization and edit helpers
- Extract _normalize_text, _build_position_map, and _find_closest_match to markdown_vault_mcp.utils.text
- Expose them as public functions (normalize_text, etc.)
- Update Collection.edit to use the new utilities
- Migrate unit tests to tests/test_text_utils.py
